requirements.txt
asgiref==3.11.0
Django==5.2.9
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
psycopg2==2.9.11
PyJWT==2.10.1
python-decouple==3.8
redis==7.1.0
sqlparse==0.5.4
typing_extensions==4.15.0
tzdata==2025.3
